lib/functions: remove redundant system commands
Tor is the primary requirement for Nipe, thus it isn't necessary to `mkdir
/etc/tor` directory in Nipe, since it's created by default on every distro
Tor is properly installed.

At the same time, the $operationalSystem var can be used to allow correct
handling of which custom Nipe's config file should be copied to /etc/tor, thus
I'm also removing the `cp` for every if statement, placing it at the end and
making use of the already available mentioned var.